refactoring & standardization: Renamed module _dio to _mds to reflect the fact that it contains the mds classes
