+++ src/astrohack/holog.py
@@ -13,7 +13,7 @@ from astrohack._utils._parm_utils._check_parms import _check_parms
 from astrohack._utils._tools import _remove_suffix
    
 from astrohack._utils._io import check_if_file_will_be_overwritten, check_if_file_exists, _read_meta_data
-from astrohack._utils._dio import AstrohackImageFile
+from astrohack._utils._mds import AstrohackImageFile
 
 def holog(
     holog_name,
