+++ src/astrohack/panel.py
@@ -12,7 +12,7 @@ from astrohack._utils._parm_utils._check_parms import _check_parms
 from astrohack._utils._tools import _remove_suffix
 from astrohack._utils._dask_graph_tools import _generate_antenna_ddi_graph_and_compute
 
-from astrohack._utils._dio import AstrohackPanelFile
+from astrohack._utils._mds import AstrohackPanelFile
 
 
 def panel(image_name, panel_name=None, cutoff=0.2, panel_model=None, panel_margins=0.2, ant_name=None, ddi=None,
